Documentation for DistCpConfiguration
